add first implementation of a grid with basic positioning
Create a grid that allows to position shapes on it.
This grid have a method to add row between elements.

Create a ShapeLayouter that produce this grid

The first implementation can only layout shapes in line

In order to test the layouting, created a SVG exporter that does a
basique SVG with shapes.