feat(converter): transform bpmn model into algorithms model (#20)
let you load a BPMN file, pass data through the algorithm, and export the result to an ASCII or SVG file.
The BPMN output export will be managed in another commit as it requires extra conversion from the internal model to BPMN.

Notice that this is the initial support for BPMN load, there is no management of lane or laneset. This should only work with processes without such elements.