add log4j2 logger along with debug logs on algo positionning
I choosed log4j2 api instead of slf4j to have the lambda parameters
available. I think it could still work with slf4j by using the
log4j-to-slf4j binding ( to be verified )
If it does not work well I'll revert that to plain slf4j api

I included the implementation of the logger directly in the app. This
might not be the case if the project is split in an app + a lib

examples of logs produce when activating debug:

```
17:44:48.669 [main] DEBUG io.process.analytics.tools.bpmn.generator.algo.ShapeLayouter - Adding step3:
Diagram:
+-------+
|step1  |
|step2  |
|step3  |
+-------+

17:44:48.671 [main] DEBUG io.process.analytics.tools.bpmn.generator.algo.ShapeLayouter - Adding step4:
Diagram:
+--------------+
|step1         |
|step2  step4  |
|step3         |
+--------------+

17:44:48.672 [main] DEBUG io.process.analytics.tools.bpmn.generator.algo.ShapeLayouter - Adding step5:
Diagram:
+--------------+
|step1         |
|step2  step4  |
|       step5  |
|step3         |
+--------------+
```