feat: improve waypoints of edges involving gateways
Merge gateways: ensure use of the bottom edge of the gateway
In particular, manage from bottom-right to top-left when top is a gateway.

Split gateways: the outgoing edges always start with a short horizontal line,
then do a polyline to the target element.
This ensure there is no overlap with incoming edges.