feat: better identify incoming and outgoing edges
Previously, it was possible to have vertically incoming and vertically outgoing edges on the same shape.
This made it difficult to distinguish one from the other.
This usually happened on gateways.

To easily identify the two edge types, only incoming edges can now be connected vertically.
Outgoing edges always start with a small horizontal segment.